Can Encrypt Files within the Same Directory
Thus far, encrypt.py can only encrypt files within the same directory as its .exe or .py file. The file scan_and_execute.py is still unfinished but intended to scan for the particular images containing the key to encrypt the files and the image containing the code to be executed.